requirements.txt
asgiref==3.3.1
certifi==2020.12.5
chardet==4.0.0
dj-database-url==0.5.0
Django==3.1.8
django-heroku==0.3.1
django-markdownx==3.0.1
django-middleware-global-request==0.2.0
gunicorn==20.0.4
idna==2.10
Markdown==3.3.3
Pillow==8.1.1
psycopg2==2.8.6
python-dateutil==1.5
pytz==2020.4
requests==2.25.1
sqlparse==0.4.1
urllib3==1.26.4
whitenoise==5.2.0
